requirements.txt
# Reddit API (GRATIS)
praw>=7.7.1

# Groq AI (GRATIS - substitui OpenAI)
groq>=0.33.0

# Google TTS (GRATIS - substitui OpenAI TTS)
gtts>=2.5.0

# Processamento de video
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9

# Imagens e legendas
Pillow>=10.0.0

# Transcrição de áudio com IA (Whisper)
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Variaveis de ambiente
python-dotenv>=1.0.0

# Requisicoes HTTP (para ElevenLabs OPCIONAL)
requests>=2.31.0
